requirements.txt
flask
pandas
matplotlib
langchain
langchain-openai
langchain-google-genai
python-dotenv
sendgrid
certifi
